Add a method to use an open filedescriptor instead of filename
This makes it possible to open a file handle in advance and operate on this
handle. This is required to ensure consistent data processing. For example, a
file handle is opened and after calculating the chunk sizes it is possible to do
a seek(0) to re-read the same data in the Python method itself.