Use unicode sample data in tests
Also ensure that the sample data used in the tests is always the
same to make test results comparable on different platforms.

Python 3 tests are also enabled on travis-ci.org now.

Note: the sample reference data changed due to the written
unicode data, but not the internal behaviour is not changed.